[worker] mcp split 3/6: extract find_impact with per-module tool_router merge
